+++ src/game/chars/CCharSpell.cpp
@@ -2980,8 +2980,8 @@ int CChar::Spell_CastStart()
 		else
 		{
 			tchar *pszTemp = Str_GetTemp();
-			size_t len = 0;
-			for ( size_t i = 0; ; i++ )
+			int len = 0;
+			for ( int i = 0; ; i++ )
 			{
 				tchar ch = pSpellDef->m_sRunes[i];
 				if ( !ch )
